REWORK CI Build:
Adapting CI's build logic to the newly reworked build logic introduced in pull request #140

+++ build/Makefiles/Docker.mk
@@ -1,21 +1,23 @@
-DOCKER_CMD ?= docker
-DOCKER_STAMP = .zig-docker
-DOCKERFILE_DIR = build/docker
+ifeq ($(DOCKER), true)
 
-$(DOCKER_STAMP): $(DOCKERFILE_DIR)/Dockerfile
-	$(DOCKER_CMD) build -t zig $(DOCKERFILE_DIR)
-	$(DOCKER_CMD) volume create --name $(VOLUME_NAME) --driver=local --opt type=none --opt device=$(PWD) --opt o=bind,uid=$(shell id -u)
-		> $(DOCKER_STAMP)
+    DOCKER_CMD ?= docker
+    DOCKER_STAMP = .zig-docker
+    DOCKERFILE_DIR = build/docker
 
-VOLUME_NAME=kfs
-ZIG = $(DOCKER_CMD) run --rm -w /build -v $(VOLUME_NAME):/build:rw zig zig
+    $(DOCKER_STAMP): $(DOCKERFILE_DIR)/Dockerfile
+		$(DOCKER_CMD) build -t zig $(DOCKERFILE_DIR)
+		$(DOCKER_CMD) volume create --name $(VOLUME_NAME) --driver=local --opt type=none --opt device=$(PWD) --opt o=bind,uid=$(shell id -u)
+			> $(DOCKER_STAMP)
 
-build: $(DOCKER_STAMP)
-debug: $(DOCKER_STAMP)
-release: $(DOCKER_STAMP)
-fast: $(DOCKER_STAMP)
-format: $(DOCKER_STAMP)
+    VOLUME_NAME=kfs
 
-.PHONY: run
-run: build
-	qemu-system-i386 -cdrom kfs.iso
+    ZIG = $(DOCKER_CMD) run --rm -w /build -v $(VOLUME_NAME):/build:rw zig zig
+
+
+    build: $(DOCKER_STAMP)
+    debug: $(DOCKER_STAMP)
+    release: $(DOCKER_STAMP)
+    fast: $(DOCKER_STAMP)
+    format: $(DOCKER_STAMP)
+
+endif
